Update Mesh.js
When normals are not provided in a geometry, they are computed for it by Mesh.calcNormals(). But this method does not take ccw field of the geometry into account. This means that geometries that do not have normals provided and have ccw=FALSE have normals of opposite direction.

This commit fixes this. Normals are computed taking ccw input parameter into account.