Update Utils.js
If there is an empty ImageTexture node (its url is empty) in x3d world, doc.downloadCount never decreases to zero in Firefox (image.onload() and/or image.onerror() does not get called). This means that progressDiv is still visible. Fix for this is obvious -- doc.downloadCount is not increased in x3dom.Utils.createTexture2D() if url is empty and texture is immediately returned instead. Solution tested in x3dom 1.5.1 only.