Don' use Number.MIN_VALUE for Vec3f.MIN
From Mozilla: "The Number.MIN_VALUE property represents the smallest
positive numeric value representable in JavaScript.

The MIN_VALUE property is the number closest to 0, not the most
negative number, that JavaScript can represent.

MIN_VALUE has a value of approximately 5e-324. Values smaller than
MIN_VALUE ("underflow values") are converted to 0."

This means: x>= MIN_VALUE == x > 0 for all x

and is almost never what one wanted it to mean. Vec3f.MIN
is used exclusively for getVolume()-related things which
suports that this is indeed -MAX_VALUE that is expected.

Signed-off-by: Simon Thum <simon.thum@igd.fraunhofer.de>