Regression Tests: Rewrote Timeline Publishing
Changed workflow of createTimelinePage to make it more roboust to changes
of the test suite. This allows to add new tests without problems caused by
missing entries in the result list of previously run tests.
Different steps of a testcase are aggregated and displayed as one test in
the timeline.