Basic DIS functionality
Dynamic creation of entities based on received DIS ESPDUs is now
working. Example scene is included in the HTML directory to demonstrate
this. It needs to be served by a host that can also forward DIS packets
over a websocket--and of course, the server must have access to a stream
of DIS packets to be forwarded.