Apply devicePixelRatio to canvas coordinates on scaled displays
Picking was broken on scaled displays, e.g. Windows scaled to 125% for high DPI display devices. This can easily be seen in e.g. the picking demo, but applied equally to any combined use of runtime.mousePosition and e.g. runtime.pickRect or accessing Event.hitPnt.  The coordinates returned for a MouseEvent did not account for the previous use of scaling on the 3d canvas element, resulting in a visible offset between the physical mouse position and the logical position computed by mousePosition.  Scaling the result of this function corrects the issue.