Increase texture coordinate precision on picking
X3DOM supports only 8-bit precision when picking texture coordinates. This is leads to poor quality results
when textures are > 256 in any dimension. The "B" channel is currently not used during this operation. This
modification uses the "B" channel to add 4 bits each to the U & V coordinates when picking textures giving a
total of 12 bits or 4096 unique coordinate values in each direction. No change was made to any computation of
S & T coordinates (X3D texture coordinates) to retain the [0.0,1.0] range. The change is in the value stored
in the picking (mouseover) event. Returned values are integers in [0,4095].