Let <code> elements wrap white space
Earlier today I tried reading your documentation page for IndexedFaceSet on my iPad. The entire page was zoomed out to accommodate the <code> element near the top, making it impossible to read without zooming in. Even for browsers that do not zoom out for extremely long lines, most of the text will be outside of the visible window and not easily accessible.

This small change keeps the HTML encoding of the node and all of its defaults in a compact wrapped area where everything is easily seen, making the documentation more useful.