Update X3DCanvas.js
The call to this._initContext with five parameters does not match the function itself, which has four parameters. This is because in the call, the parameter for "flashie" is pased as the fourth parameter, while in the function itself this parameter has been removed. As a consequence, one could not use the tag <X3D backend="webgl2"> to invoke a webgl2 context, because the 'tryWebGL2' parameter is lost as the fifth parameter. The proposed change fixes this error.