Background Cube
Background Cube images look different between x3dom 1.7.2 (good) and  1.8.1 (ugly). The reason was not as initially suspected, the size of the generated background sphere - this size is in fact irrelevant- but because the wrong shader name BACKGROUND_CUBETEXTURE_DDS which is used even if you do not provide an dds file.

After replacing the shader name, the images look as expected and all images are connected correctly to fill the cube.
However, there is still one more difference: the view is flipped upside down.

So far I tested only using http://github/x3d-solar-system with the original images and another background picture set.